// frontend/src/app/api/model/defaultView.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const DefaultView = {
    Dashboard: 'DASHBOARD',
    Tasks: 'TASKS',
    Habits: 'HABITS',
    Learning: 'LEARNING',
    Plans: 'PLANS',
} as const;
export type DefaultView = typeof DefaultView[keyof typeof DefaultView];